Make WorkflowStepResolution an entity
This way we'll be able to create a view for it, so the user
can follow its progress and edit it manually.
